Remove duplicated code between Driver extraction classes
