+++ src/psyclone/domain/lfric/lfric_extract_driver_creator.py
@@ -52,16 +52,13 @@ from psyclone.parse import ModuleManager
 from psyclone.psyGen import InvokeSchedule
 from psyclone.psyir.backend.fortran import FortranWriter
 from psyclone.psyir.frontend.fortran import FortranReader
-from psyclone.psyir.nodes import (Assignment, Call, FileContainer,
-                                  IntrinsicCall, Literal, Reference,
-                                  Routine, ExtractNode, StructureReference)
-from psyclone.psyir.symbols import (ArrayType, CHARACTER_TYPE,
+from psyclone.psyir.nodes import (Call, FileContainer,
+                                  Routine, StructureReference)
+from psyclone.psyir.symbols import (
                                     ContainerSymbol, DataSymbol,
                                     DataTypeSymbol, UnresolvedType,
                                     ImportInterface, INTEGER_TYPE,
-                                    RoutineSymbol, UnsupportedFortranType,
-                                    AutomaticInterface)
-from psyclone.psyir.transformations import ExtractTrans
+                                    RoutineSymbol, UnsupportedFortranType)
 
 
 class LFRicExtractDriverCreator(BaseDriverCreator):
@@ -195,236 +192,6 @@ class LFRicExtractDriverCreator(BaseDriverCreator):
 
         '''
         return name.replace("-", "")[:63]
-
-    # -------------------------------------------------------------------------
-    @staticmethod
-    def _create_output_var_code(name, program, is_input, read_var,
-                                postfix, index=None, module_name=None):
-        # pylint: disable=too-many-arguments
-        '''
-        This function creates all code required for an output variable.
-        It creates the '_post' variable which stores the correct result
-        from the file, which is read in. If the variable is not also an
-        input variable, the variable itself will be declared (based on
-        the size of the _post variable) and initialised to 0.
-        This function also handles array of fields, which need to get
-        an index number added.
-        If a module_name is specified, this indicates that this variable
-        is imported from an external module. The name of the module will
-        be appended to the tag used in the extracted kernel file, e.g.
-        `dummy_var2@dummy_mod`.
-
-        :param str name: the name of original variable (i.e.
-            without _post), which will be looked up as a tag in the symbol
-            table. If index is provided, it is incorporated in the tag using
-            f"{name}_{index}_data".
-        :param program: the PSyIR Routine to which any code must
-            be added. It also contains the symbol table to be used.
-        :type program: :py:class:`psyclone.psyir.nodes.Routine`
-        :param bool is_input: True if this variable is also an input
-            parameter.
-        :param str read_var: the readvar method to be used including the
-            name of the PSyData object (e.g. 'psy_data%ReadVar')
-        :param str postfix: the postfix to use for the expected output
-            values, which are read from the file.
-        :param index: if present, the index to the component of a field vector.
-        :type index: Optional[int]
-        :param str module_name: if the variable is part of an external module,
-            this contains the module name from which it is imported.
-            Otherwise, this must either not be specified or an empty string.
-
-        :returns: a 2-tuple containing the output Symbol after the kernel,
-             and the expected output read from the file.
-        :rtype: Tuple[:py:class:`psyclone.psyir.symbols.Symbol`,
-                      :py:class:`psyclone.psyir.symbols.Symbol`]
-
-        '''
-        # For each variable that is written, we need to declare a new variable
-        # that stores the expected value which is contained in the kernel data
-        # file, which has `_post` appended to the name (so `a` is the variable
-        # that is written, and `a_post` is the corresponding variable that
-        # has the expected results for verification). Since the written
-        # variable and the one storing the expected results have the same
-        # type, look up the 'original' variable and declare the _POST variable
-        symbol_table = program.symbol_table
-        if module_name:
-            sym = symbol_table.lookup_with_tag(f"{name}@{module_name}")
-        else:
-            sym = symbol_table.lookup(name)
-
-        # Declare a 'post' variable of the same type and read in its value.
-        post_name = sym.name + postfix
-        post_sym = symbol_table.new_symbol(post_name,
-                                           symbol_type=DataSymbol,
-                                           datatype=sym.datatype.copy())
-        if isinstance(post_sym.datatype, UnsupportedFortranType):
-            post_sym.datatype = post_sym.datatype.copy()
-            post_sym.datatype._declaration = \
-                post_sym.datatype._declaration.replace(sym.name, post_name)
-
-        if module_name:
-            post_tag = f"{name}{postfix}@{module_name}"
-        else:
-            if index is not None:
-                post_tag = f"{name}{postfix}%{index}"
-            else:
-                # If it is not indexed then `name` will already end in "_data"
-                post_tag = f"{name}{postfix}"
-        name_lit = Literal(post_tag, CHARACTER_TYPE)
-
-        if sym.is_array and not sym.datatype.is_allocatable:
-            # In case of a non-allocatable array (e.g. a constant
-            # size array from a module), call the ReadVariable
-            # function that does not require an allocatable field
-            BaseDriverCreator.add_call(program, read_var+"NonAlloc",
-                                       [name_lit, Reference(post_sym)])
-        else:
-            # In case of an allocatable array, call the ReadVariable
-            # function that will also allocate this array.
-            BaseDriverCreator.add_call(program, read_var,
-                                       [name_lit, Reference(post_sym)])
-
-        # Now if a variable is written to, but not read, the variable
-        # is not allocated. So we need to allocate it and set it to 0.
-        if not is_input:
-            if (isinstance(post_sym.datatype, ArrayType) or
-                    (isinstance(post_sym.datatype, UnsupportedFortranType) and
-                     isinstance(post_sym.datatype.partial_datatype,
-                                ArrayType))):
-                alloc = IntrinsicCall.create(
-                    IntrinsicCall.Intrinsic.ALLOCATE,
-                    [Reference(sym), ("mold", Reference(post_sym))])
-                program.addchild(alloc)
-            set_zero = Assignment.create(Reference(sym),
-                                         Literal("0", INTEGER_TYPE))
-            program.addchild(set_zero)
-        return (sym, post_sym)
-
-    # -------------------------------------------------------------------------
-    def _create_read_in_code(self, program, psy_data, original_symbol_table,
-                             read_write_info, postfix):
-        # pylint: disable=too-many-arguments, too-many-branches
-        # pylint: disable=too-many-locals, too-many-statements
-        '''This function creates the code that reads in the NetCDF file
-        produced during extraction. For each:
-
-        - variable that is read-only, it will declare the symbol and add code
-          that reads in the variable using the PSyData library.
-        - variable that is read and written, it will create code to read in the
-          variable that is read, and create a new variable with the same name
-          and "_post" added which is read in to store the values from the
-          NetCDF file after the instrumented region was executed. In the end,
-          the variable that was read and written should have the same value
-          as the corresponding "_post" variable.
-        - variable that is written only, it will create a variable with "_post"
-          as postfix that reads in the output data from the NetCDF file. It
-          then also declares a variable without postfix (which will be the
-          parameter to the function), allocates it based on the shape of
-          the corresponding "_post" variable, and initialises it with 0.
-
-        :param program: the PSyIR Routine to which any code must
-            be added. It also contains the symbol table to be used.
-        :type program: :py:class:`psyclone.psyir.nodes.Routine`
-        :param psy_data: the PSyData symbol to be used.
-        :type psy_data: :py:class:`psyclone.psyir.symbols.DataSymbol`
-        :param read_write_info: information about all input and output
-            parameters.
-        :type read_write_info: :py:class:`psyclone.psyir.tools.ReadWriteInfo`
-        :param str postfix: a postfix that is added to a variable name to
-            create the corresponding variable that stores the output
-            value from the kernel data file.
-
-        :returns: all output parameters, i.e. variables that need to be
-            verified after executing the kernel. Each entry is a 2-tuple
-            containing the symbol of the computed variable, and the symbol
-            of the variable that contains the value read from the file.
-        :rtype: List[Tuple[:py:class:`psyclone.psyir.symbols.Symbol`,
-                           :py:class:`psyclone.psyir.symbols.Symbol`]]
-
-        '''
-        symbol_table = program.scope.symbol_table
-        read_var = f"{psy_data.name}%ReadVariable"
-        mod_man = ModuleManager.get()
-
-        # First handle variables that are read:
-        # -------------------------------------
-        read_stmts = []
-        for module_name, signature in read_write_info.read_list:
-            if module_name:
-                continue
-            orig_sym = original_symbol_table.lookup(signature[0])
-            sym = orig_sym.copy()
-            sym.interface = AutomaticInterface()
-            if symbol_table.lookup(sym.name, otherwise=None) is not None:
-                # We can edit the name because we know the copied symbol is
-                # not in a symbol table yet
-                sym._name = symbol_table.next_available_name(sym.name)
-            symbol_table.add(sym)
-            name_lit = Literal(str(signature), CHARACTER_TYPE)
-            read_stmts.append((name_lit, sym))
-
-        ExtractNode._bring_external_symbols(read_write_info,
-                                            program.scope.symbol_table)
-        for module_name, signature in read_write_info.read_list:
-            if not module_name:
-                continue
-            mod_info = mod_man.get_module_info(module_name)
-            orig_sym = mod_info.get_symbol(signature[0])
-            tag = f"{signature[0]}@{module_name}"
-            sym = symbol_table.lookup_with_tag(tag)
-            name_lit = Literal(tag, CHARACTER_TYPE)
-            read_stmts.append((name_lit, sym))
-
-        for name_lit, sym in read_stmts:
-            # TODO #2898: the test for array can be removed if
-            # `is_allocatable` is supported for non-arrays.
-            if sym.is_array and not sym.datatype.is_allocatable:
-                # In case of a non-allocatable array (e.g. a constant
-                # size array from a module), call the ReadVariable
-                # function that does not require an allocatable field
-                self.add_call(program, read_var+"NonAlloc",
-                              [name_lit, Reference(sym)])
-            else:
-                # In case of an allocatable array, call the ReadVariable
-                # function that will also allocate this array.
-                self.add_call(program, read_var,
-                              [name_lit, Reference(sym)])
-
-        # Then handle all variables that are written (note that some
-        # variables might be read and written)
-        # ----------------------------------------------------------
-        # Collect all output symbols to later create the tests for
-        # correctness. This list stores 2-tuples: first one the
-        # variable that stores the output from the kernel, the second
-        # one the variable that stores the output values read from the
-        # file. The content of these two variables should be identical
-        # at the end.
-        output_symbols = []
-
-        for module_name, signature in read_write_info.write_list:
-            # Find the right symbol for the variable. Note that all variables
-            # in the input and output list have been detected as being used
-            # when the variable accesses were analysed. Therefore, these
-            # variables have References, and will already have been declared
-            # in the symbol table (in _add_all_kernel_symbols).
-            if module_name:
-                orig_sym = mod_man.get_module_info(module_name).get_symbol(
-                    signature[0])
-                if not orig_sym:
-                    # TODO 2120: We likely couldn't parse the module.
-                    print(f"Error finding symbol '{signature}' in "
-                          f"'{module_name}'.")
-                    continue
-            else:
-                orig_sym = original_symbol_table.lookup(signature[0])
-            is_input = read_write_info.is_read(signature)
-            sym_tuple = \
-                self._create_output_var_code(str(signature), program,
-                                             is_input, read_var, postfix,
-                                             module_name=module_name)
-            output_symbols.append(sym_tuple)
-
-        return output_symbols
 
     # -------------------------------------------------------------------------
     @staticmethod
@@ -603,14 +370,6 @@ class LFRicExtractDriverCreator(BaseDriverCreator):
 
         '''
         # pylint: disable=too-many-locals
-
-        # Since this is a 'public' method of an entirely separate class,
-        # we check that the list of nodes is what it expects. This is done
-        # by invoking the validate function of the basic extract function.
-        extract_trans = ExtractTrans()
-        # We need to provide the prefix to the validation function:
-        extract_trans.validate(nodes, options={"prefix": prefix})
-
         module_name, local_name = region_name
         unit_name = self._make_valid_unit_name(f"{module_name}_{local_name}")
